backend/pyproject.toml
[tool.poetry]
name = "dreamdrugs"
version = "0.1.0"
description = "Deep dreaming on a Raspberry Pi"
authors = ["Ramon Blanquer <blanquer.ramon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
gunicorn = "^19.9"
numpy = "^1.17"
redis = "^3.3"
scipy = "^1.3"
Flask = "^1.1"
Pillow = "^6.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.7"
pylint = "^2.4"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest-cov = "^2.8"
remote-pdb = "^2.0.0"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
